Cleaned up code and separated AuthPreference into its own file.
