+++ app/src/main/java/org/geysermc/app/android/ui/geyser/user_auths/AuthPreference.java
@@ -1,0 +1,63 @@
+/*
+ * Copyright (c) 2020-2020 GeyserMC. http://geysermc.org
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ *
+ * @author GeyserMC
+ * @link https://github.com/GeyserMC/GeyserAndroid
+ */
+
+package org.geysermc.app.android.ui.geyser.user_auths;
+
+import android.content.Context;
+
+import androidx.preference.Preference;
+import androidx.preference.PreferenceViewHolder;
+
+/**
+ * Implement custom androidx.preference.Preference with long press support.
+ */
+class AuthPreference extends Preference
+{
+    private OnHoldListener onHoldListener;
+
+    public AuthPreference(Context context) {
+        super(context);
+    }
+
+    @Override
+    public void onBindViewHolder(PreferenceViewHolder holder) {
+        super.onBindViewHolder(holder);
+        holder.itemView.setOnLongClickListener(view -> {
+            onHoldListener.onLongClick(this);
+            return true;
+        });
+    }
+
+    public void setOnHoldListener(OnHoldListener onHoldListener) {
+        this.onHoldListener = onHoldListener;
+    }
+
+    /**
+     * Implement OnHoldListener interface.
+     */
+    interface OnHoldListener {
+        void onLongClick(AuthPreference preference);
+    }
+}
