+++ app/src/main/res/values/strings.xml
@@ -89,17 +89,15 @@
     <string name="user_auths_save_save">Save</string>
     <string name="user_auths_save_failed">Unable to save user auths!</string>
     <string name="user_auths_save_discard">Discard</string>
+    <string name="user_auths_delete_dialog_title">User Auth</string>
+    <string name="user_auths_delete_dialog_message">Are you sure you\'d like to delete the selected entry?</string>
+    <string name="user_auths_delete_dialog_positive">Delete</string>
+    <string name="user_auths_delete_dialog_negative">Cancel</string>
 
     <!-- User Auth Dialog -->
     <string name="user_auth_dialog_title">User Auth</string>
     <string name="user_auth_dialog_positive">Save</string>
     <string name="user_auth_dialog_negative">Cancel</string>
-
-    <!-- User Auth Delete Confirmation Dialog -->
-    <string name="user_auth_delete_dialog_title">User Auth</string>
-    <string name="user_auth_delete_dialog_message">Are you sure you\'d like to delete the selected entry?</string>
-    <string name="user_auth_delete_dialog_positive">Delete</string>
-    <string name="user_auth_delete_dialog_negative">Cancel</string>
 
     <!-- About Screen -->
     <string name="about_about_title">About</string>
